+++ src/firmware/boards/pico.h
@@ -1,1 +1,0 @@
-quokkadb.h