+++ src/firmware/lib/QuokkADB/include/adb_platform.h
@@ -1,6 +1,6 @@
 
 //---------------------------------------------------------------------------
-//  BT-USB-ADB-Adapter
+//  ultrausbt-Apple-ADB-adapter
 //  This project is based on adbuino and QuokkADB:
 //
 //	ADBuino & QuokkaADB ADB keyboard and mouse adapter