+++ src/firmware/lib/QuokkADB/src/adb_platform.cpp
@@ -1,5 +1,5 @@
 //---------------------------------------------------------------------------
-//  BT-USB-ADB-Adapter
+//  ultrausbt-Apple-ADB-adapter
 //  This project is based on adbuino and QuokkaADB:
 //
 //	QuokkaADB ADB keyboard and mouse adapter