+++ src/firmware/lib/QuokkADB/src/platformkbdparser.cpp
@@ -1,5 +1,5 @@
 //----------------------------------------------------------------------------
-//  BT-USB-ADB-Adapter
+//  ultrausbt-Apple-ADB-adapter
 //  This project is based on adbuino and QuokkADB:
 //
 //  QuokkADB ADB keyboard and mouse adapter