+++ src/firmware/lib/usb/src/char2usbkeycode.cpp
@@ -1,5 +1,5 @@
 //---------------------------------------------------------------------------
-//  BT-USB-ADB-Adapter
+//  ultrausbt-Apple-ADB-adapter
 //  This project is based on adbuino and QuokkADB:
 //
 //	ADBuino & QuokkaADB ADB keyboard and mouse adapter