+++ src/firmware/src/sdkconfig.h
@@ -1,5 +1,5 @@
 //
-// Bluepad32 SDK configuration for BT-USB-ADB-Adapter
+// Bluepad32 SDK configuration for ultrausbt-Apple-ADB-adapter
 //
 #define CONFIG_BLUEPAD32_MAX_DEVICES 4
 #define CONFIG_BLUEPAD32_MAX_ALLOWLIST 4